Add saml to docker-compose
To get it working, went back to a simpler setup without
SSL and proxy. Also removed some of the services I did not
need to test this. It took a lot of trial and error to get
this working. I am not sure networking was ever a factor
in the problems but doing it this way just made things
easier since the back channel and front channel networks
could be the same hostnames and ports.